requirements.txt
django>=1.9
django-braces
djangorestframework
django-extensions
django-test-without-migrations
pygments
markdown
BeautifulSoup
pyyaml
pytz
six
lxml
